# value-stream-service/src/api-v1/swagger.yaml
swagger: '2.0'
info:
  description: Value Stream Map Service
  version: 1.0.0
  title: Value Stream Map Service

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
 
paths:
  /:
    get:
      summary: retrieves API
      operationId: getAPI
      produces:
        - application/json
        - text/html
      responses:
        200:
          description: Found
  /valuestreams:
    get:
      summary: retrieves value streams
      operationId: getValueStreams
      description: Return all value stream matching the search criteria
      produces:
        - application/json
        - text/html
      parameters:
        - in: query
          name: teamName
          description: Team name fragment to find
          type: string
          x-express-openapi-case-sensitive: false
          required: false
        - in: query
          name: teamId
          description: Team ID to find
          type: string
          x-express-openapi-case-sensitive: false
          required: false
      responses:
        200:
          description: search results matching criteria
          schema:
            type: array
            items:
              $ref: '#/definitions/ValueStream'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: adds a value stream
      operationId: addValueStream
      description: Adds a value stream
      consumes:
        - application/json
      produces:
        - application/json
        - text/html
      parameters:
        - in: body
          name: valueStream
          description: Value stream to add
          required: true
          schema:
            $ref: '#/definitions/ValueStream'
      responses:
        201:
          description: Value stream created
          schema:
            $ref: '#/definitions/ValueStream'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  /valuestreams/{teamId}:
    get:
      summary: Gets a value stream by ID.
      description: Returns a value stream for a team
      operationId: getValueStream
      produces:
        - application/json
        - text/html
      parameters:
        - in: path
          name: teamId
          description: Team ID
          type: string
          x-express-openapi-case-sensitive: false
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ValueStream'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  /valuestreams/{teamId}/process:
    put:
      summary: Adds a process to a value stream
      operationId: addProcess
      description: Adds a process to a value stream
      consumes:
        - application/json
      produces:
        - application/json
        - text/html
      parameters:
        - in: path
          name: teamId
          description: Team ID
          type: string
          x-express-openapi-case-sensitive: false
          required: true
        - in: body
          name: process
          description: Process to add
          required: true
          schema:
            $ref: '#/definitions/Process'
      responses:
        201:
          description: Process added
          schema:
            $ref: '#/definitions/ValueStream'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

x-swagger-mongoose:
  schema-options:
    timestamps: true

definitions:
  ValueStream:
    type: object
    required:
      - teamName
      - teamId
      - status
    example:
      teamName: X Force
      teamId: TEAM12345
      status: ACTIVE
      processes:
        - stepId: STEP12
          name: Coding
          description: Writing the code
          reprocessTime: 20
          processTime: 80
          waitTime: 120
          primaryPath:
            nextProcess: STEP13
            frequencyPct: 100        
        - stepId: STEP13
          name: Code Review
          description: Validating tests
          reprocessTime: 20
          processTime: 80
          waitTime: 120
          primaryPath:
            nextProcess: STEP14
            frequencyPct: 80
          exceptionPath:
            nextProcess: STEP12
            frequencyPct: 20
    properties:
      teamId:
        type: string
        minLength: 1
      teamName:
        type: string
        minLength: 1
      status:
        type: string
        description: ACTIVE or DELETE
      processes:
        type: array
        items: 
          $ref: '#/definitions/Process'
  Process:
    required:
      - stepId
      - name
      - processTime
      - waitTime
      - primaryPath
    example:
      stepId: STEP12 
      name: Coding
      description: Writing the code
      reprocessTime: 20
      processTime: 80
      waitTime: 120
      primaryPath:
        nextProcess: theNextProcessStep
        frequencyPct: 80
    properties:
      stepId:
        type: string
      name:
        type: string
        minLength: 1
        description: The name of the process step
      description:
        type: string
        description: Any clarification of the step
      reprocessTime:
        type: integer
        minimum: 0
        description: The time it takes to fix defects in the process
      processTime:
        type: integer
        minimum: 0
        description: The number of minutes spent working on the process
      waitTime:
        type: integer
        minimum: 0
        description: The number of minutes spent waiting on the process to start
      primaryPath:
        $ref: '#/definitions/Path'
      exceptionPath:
        $ref: '#/definitions/Path'
  Path:
    required:
      - nextProcess
      - frequencyPct
    properties:
      nextProcess:
        type: string
        description: The ID of the next process step
      frequencyPct:
        type: integer
        minimum: 0
        maximum: 100
        description: Percentage this path is takem
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string

schemes:
 - http
 - https

basePath: /v1